+++ src/num/float/wrapper/shared_series.rs
@@ -24,7 +24,7 @@ macro_rules! impl_float_shared_series {
     };
     (@$f:ty:$uf:ty, $ue:ty) => {
         ///
-        /// # *Common methods with or without `std` or `libm`*.
+        /// # *Common methods with or without `std`*.
         ///   *Implemented using Taylor series.*
         impl Float<$f> {
             /// Raises itself to the `y` floating point power using the Taylor series via the